LicenseRenew/LicenseRenew/urls.py
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from django.contrib import admin
from django.urls import include, path
from django.conf import settings
from django.conf.urls.static import static
from django.contrib.auth import views as auth_views
from django.http import HttpResponse


def home_view(request):
    return HttpResponse("Welcome to License Renew API")


urlpatterns = [
    
    path('', home_view, name='home'),

    path('admin/', admin.site.urls),

    path('api/', include('LicenseTrack.urls')),

      # JWT Authentication endpoints
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
   

]



if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)


