LicenseRenew/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.1
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
cron-descriptor==1.4.5
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
Django==5.2.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-writable-nested==0.7.2
fonttools==4.58.1
fuzzywuzzy==0.18.0
gunicorn
kiwisolver==1.4.8
kombu==5.5.4
Levenshtein==0.27.1
matplotlib==3.10.3
numpy==2.2.6
odfpy==1.4.1
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pymemcache==4.0.0
PyMuPDF==1.26.0
pyparsing==3.2.3
pyphen==0.17.2
pytesseract==0.3.13
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.2
RapidFuzz==3.13.0
redis==6.2.0
reportlab==4.4.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
